feat(runner): add matrix, if conditions, continue-on-error, and timeout
- Implement matrix expansion with Cartesian product, include/exclude
- Add matrix context interpolation (${{ matrix.<key> }})
- Implement step-level and job-level if conditions with status functions
- Add success(), failure(), always(), cancelled() status functions
- Implement continue-on-error with outcome/conclusion semantics
- Add timeout-minutes support for steps and jobs (float64 support)
- Extend expression interpolation for matrix context
- Update dry-run to show expanded matrix jobs
- Add test fixtures for all new features
- All tests pass: go build, go vet, go test ./...

+++ internal/workflow/model.go
@@ -19,7 +19,7 @@ type Job struct {
 	Steps           []Step            `yaml:"steps"`
 	Strategy        *Strategy         `yaml:"strategy"`
 	Outputs         map[string]string `yaml:"outputs"`
-	TimeoutMinutes  int               `yaml:"timeout-minutes"`
+	TimeoutMinutes  float64           `yaml:"timeout-minutes"`
 	Defaults        *Defaults         `yaml:"defaults"`
 	// Runtime fields
 	instanceID      string
@@ -57,5 +57,5 @@ type Step struct {
 	Shell             string         `yaml:"shell"`
 	WorkingDirectory  string         `yaml:"working-directory"`
 	ContinueOnError   bool           `yaml:"continue-on-error"`
-	TimeoutMinutes    int            `yaml:"timeout-minutes"`
+	TimeoutMinutes    float64        `yaml:"timeout-minutes"`
 }
